alx: Unified OS driver attempt for Linux / FreeBSD
Goal is to use modern tools like coccinelle to see if we can
unify the alx Ethernet driver between Linux and FreeBSD.
The objective is to see if in practice even with these
tools we can keep getting updates to both OSes through
a single tree but at the same time see what challenges
we face.

This is part of the lofty effort on killing proprietary drivers.
I have taken the alx driver code from public postings on public
mailing lists. At this point I have not updated the alx driver
to match the newest v2 patch iterations. Work is still being
done to modify that driver to meet community standards.

Ideally we should try to first complete the driver with FreeBSD
support here prior to getting it upstream in the Linux kernel
as it would make the development easier. Once we have support
targets for both FreeBSD and Linux we can push upstream.

If we don't care then forget about this project.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>