alx: update linux target
compat-drivers had removed alx from being cherry picked
as a crap type patch. The alx driver is not yet upstream
and a compromise has been reached to try to help enable
users with the alx module and to help address killing
proprietary drivers. This unified alx git tree will be
used to supply the target OS support for Linux and FreeBSD,
and compat-drivers will cherry pick the alx driver from
here while it is not yet upstream as the project aims to
provide a foundation for how to do open unified driver
development.

This also adds the -DCOMPAT_BASE which is now required
on the latest compat-drivers.

The changes in this commit address the latest changes
to compat-drivers to support unified drivers.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>