alx: add LICENSE and extend documentation for development process
The LICENSE is ISC and some original code came from me based
on the compat project, I have relicensed that stuff to ISC
and contributed here. The rest is the QCA alx driver which
is all ISC licensed.

The documentation for submitting patches and using the
Signed-off-by is added. This follows the Linux kernel's
Developer's Certificate of Origin 1.1. We don't yet have
a mailing list but what we do have is commitment to support
this by the QCA alx development team. Adrian has offered
to help address the BSD target support.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>